updated  navbar
