+++ frontend/package.json
@@ -15,6 +15,7 @@
     "mongoose": "^6.3.5",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-icons": "^4.8.0",
     "react-router-dom": "^6.2.1",
     "react-scripts": "^5.0.0",
     "web-vitals": "^2.1.4"
