--- frontend/src/components/Navbar.js
@@ -1,132 +1,0 @@
-import React from "react";
-import { Link, useNavigate, useLocation } from "react-router-dom";
-
-function Navbar() {
-  const navigate = useNavigate();
-  const location = useLocation();
-  const handleLogout = () => {
-    localStorage.removeItem("token");
-    navigate("/login");
-  };
-
-  return (
-    <div>
-      <nav
-        className="navbar navbar-expand-lg navbar-light"
-        style={{ backgroundColor: "#e3f2fd", opacity: "70%" }}
-      >
-        <div className="container-fluid">
-          <a className="navbar-brand" href="#">
-            NOTEBOOK
-          </a>
-          <button
-            className="navbar-toggler"
-            type="button"
-            data-bs-toggle="collapse"
-            data-bs-target="#navbarSupportedContent"
-            aria-controls="navbarSupportedContent"
-            aria-expanded="false"
-            aria-label="Toggle navigation"
-          >
-            <span className="navbar-toggler-icon"></span>
-          </button>
-          <div className="collapse navbar-collapse" id="navbarSupportedContent">
-            <ul className="navbar-nav me-auto mb-2 mb-lg-0">
-              <li className="nav-item">
-                <Link
-                  to="/notes"
-                  className={`nav-link ${
-                    location.pathname === "/notes" ? "active" : ""
-                  }`}
-                  aria-current="page"
-                >
-                  Your Notes
-                </Link>
-              </li>
-
-              <li className="nav-item">
-                <Link
-                  to="/"
-                  className={`nav-link ${
-                    location.pathname === "/" ? "active" : ""
-                  }`}
-                  aria-current="page"
-                >
-                  AddNote
-                </Link>
-              </li>
-
-              <li className="nav-item">
-                <Link
-                  to="/about"
-                  className={`nav-link ${
-                    location.pathname === "/about" ? "active" : ""
-                  }`}
-                >
-                  About
-                </Link>
-              </li>
-            </ul>
-
-            {!localStorage.getItem("token") ? (
-              <form className="d-flex">
-                <Link
-                  to="/login"
-                  className={`nav-link ${
-                    location.pathname === "/login" ? "active" : ""
-                  }`}
-                >
-                  Login
-                </Link>
-
-                <Link
-                  to="/signup"
-                  className={`nav-link ${
-                    location.pathname === "/signup" ? "active" : ""
-                  }`}
-                >
-                  SignUp
-                </Link>
-              </form>
-            ) : (
-              <>
-                
-                {localStorage.getItem("name") ? (
-                  <div className="h3 px-10">{localStorage.getItem("name")}</div>
-                ) : (
-                  ""
-                )}<Link
-                to="/login"
-                onClick={handleLogout}
-                className={`nav-link px-3 ${
-                  location.pathname === "/logout" ? "active" : ""
-                }`}
-              >
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  width="25"
-                  height="25"
-                  fill="currentColor"
-                  class="bi bi-box-arrow-right h3"
-                  viewBox="0 0 16 16"
-                >
-                  <path
-                    fill-rule="evenodd"
-                    d="M10 12.5a.5.5 0 0 1-.5.5h-8a.5.5 0 0 1-.5-.5v-9a.5.5 0 0 1 .5-.5h8a.5.5 0 0 1 .5.5v2a.5.5 0 0 0 1 0v-2A1.5 1.5 0 0 0 9.5 2h-8A1.5 1.5 0 0 0 0 3.5v9A1.5 1.5 0 0 0 1.5 14h8a1.5 1.5 0 0 0 1.5-1.5v-2a.5.5 0 0 0-1 0v2z"
-                  />
-                  <path
-                    fill-rule="evenodd"
-                    d="M15.854 8.354a.5.5 0 0 0 0-.708l-3-3a.5.5 0 0 0-.708.708L14.293 7.5H5.5a.5.5 0 0 0 0 1h8.793l-2.147 2.146a.5.5 0 0 0 .708.708l3-3z"
-                  />
-                </svg>
-              </Link>
-
-              </>
-            )}
-          </div>
-        </div>
-      </nav>
-    </div>
-  );
-}
-export default Navbar;
