+++ frontend/src/components/componentStyle/Navbar.css
@@ -1,0 +1,160 @@
+@import url("https://fonts.googleapis.com/css2?family=Titillium+Web:wght@300;400;700&display=swap");
+
+* {
+	padding: 0;
+	margin: 0;
+	box-sizing: border-box;
+}
+
+:root {
+	--mainColor: #29335c;
+	--mainColorLight: #5767aa;
+	--secondaryColor: #db2b39;
+	--textColor: #eee;
+}
+
+header {
+	display: flex;
+	align-items: center;
+  position: sticky;
+  top:0;
+  align-items: center;
+   font-family: 'Poppins', sans-serif;
+   font-weight: bold;
+   letter-spacing: 2px;
+	justify-content: space-between;
+	height: 80px;
+	padding: 0 2rem;
+	background-color: var(--mainColor);
+	color: var(--textColor);
+}
+
+header .logo{
+	/* margin-bottom:-0.0px; */
+	margin-bottom:-0px;
+	margin-right:3px;
+}
+
+ .authbuttons{
+  display: flex;
+  column-gap: 5px;
+}
+.auth-btn{
+	width:100px;
+padding:5px 10px;
+white-space: nowrap;
+
+border-radius: 10px;
+font-weight:bold;
+background-color: #FFC40C;
+
+}
+
+.auth{
+	width:75px;
+}
+.auth-btn a{
+	color:black;
+	text-decoration: none;
+}
+
+nav a {
+	margin: 0 1rem;
+	color: var(--textColor);
+	text-decoration: none;
+}
+
+
+.userIcon{
+	margin-right:4px;
+	font-size:40px;;
+}
+
+header .nav-btn {
+	cursor: pointer;
+	background: transparent;
+	border: none;
+	outline: none;
+	color: var(--textColor);
+	display: none;
+	opacity: 0;
+	font-size: 1.8rem;
+	margin-left:5px;
+}
+
+header div,
+nav {
+	display: flex;
+	align-items: center;
+}
+
+
+@media only screen and (max-width: 1024px) {
+	header .nav-btn {
+		display: block;
+		opacity: 1;
+	}
+
+	header nav {
+		position: fixed;
+		top: -100vh;
+		left: 0;
+		height: 100%;
+		width: 100%;
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+		justify-content: center;
+		gap: 1.5rem;
+		background-color: var(--mainColor);
+		transition: 1s;
+	}
+
+	header .responsive_nav {
+		transform: translateY(100vh);
+	}
+
+	nav .nav-close-btn {
+		position: absolute;
+		top: 2rem;
+		right: 2rem;
+	}
+
+	nav a {
+		font-size: 1.5rem;
+	}
+}
+
+@media screen and (max-width:400px){
+	header{
+		padding:2rem 1rem;
+	}
+
+}
+
+
+/* underline hover effects */
+a.underline-hover-effect {
+  text-decoration: none;
+  color: inherit;
+}
+
+.underline-hover-effect {
+  display: inline-block;
+  padding-bottom: 0.25rem; /* defines the space between text and underline */
+  position: relative;
+}
+
+.underline-hover-effect::before {
+  content: "";
+  position: absolute;
+  left: 0;
+  bottom: 0;
+  width: 0;
+  height: 2px;
+  background-color: #FFC40C;
+  transition: width 0.25s ease-out;
+}
+.underline-hover-effect:hover::before {
+  width: 100%;
+}
